+++ libvirt/tests/src/virsh_cmd/domain/virsh_destroy.py
@@ -52,14 +52,18 @@ def run(test, params, env):
     elif vm_ref == "uuid":
         vm_ref = domuuid
 
+    uri = params.get("connect_uri")
+    if vm_ref.find("--graceful") > 0 and uri == "lxc:///":
+        raise error.TestNAError("lxc does not support --graceful")
+
     if libvirtd == "off":
         utils_libvirtd.libvirtd_stop()
 
     if vm_ref != "remote":
-        status = virsh.destroy(vm_ref, ignore_status=True,
+        output = virsh.destroy(vm_ref, ignore_status=True,
                                unprivileged_user=unprivileged_user,
-                               uri=uri, debug=True).exit_status
-        output = ""
+                               uri=uri, debug=True)
+        status = output.exit_status
     else:
         status = 0
         try:
@@ -76,6 +80,9 @@ def run(test, params, env):
 
     if libvirtd == "off":
         utils_libvirtd.libvirtd_start()
+
+    if vm.state() == "paused":
+        vm.resume()
 
     # check status_error
     if status_error == "yes":
